fix(batch): filter saved videos by status 'ENDED' (#5872)
This pull request introduces a small but important change to the
`getSavedVideos` function. Now, only videos with a `status` of `'ENDED'`
will be returned, in addition to those not marked as deleted. This
ensures that only completed videos are included in the results.

* Restricts the query in `getSavedVideos`
(`apps/batch/app/videos/check/_lib/get-saved-videos.ts`) to return only
videos where `status` is `'ENDED'`.